Move first function from Covid to sm_sir
