+++ autumn/models/covid_19/detection.py
@@ -1,4 +1,4 @@
-from typing import Callable, Optional, Tuple, Any, List
+from typing import Callable, Optional, Tuple, List
 import numpy as np
 
 from summer.compute import ComputedValueProcessor
@@ -14,10 +14,7 @@ from autumn.core.inputs.covid_lka.queries import get_lka_testing_numbers
 from autumn.core.inputs.covid_mmr.queries import get_mmr_testing_numbers
 from autumn.core.inputs.covid_bgd.queries import get_coxs_bazar_testing_numbers
 from autumn.core.inputs.owid.queries import get_international_testing_numbers
-from autumn.core.inputs import get_population_by_agegroup
-from autumn.core.utils.utils import apply_moving_average
 from autumn.model_features.curve import scale_up_function
-from autumn.models.covid_19.stratifications.agegroup import AGEGROUP_STRATA
 
 
 class CdrProc(ComputedValueProcessor):
@@ -131,59 +128,3 @@ def inflate_test_data(
     return [
         test_values[val] * testing_scale_up(time) for val, time in enumerate(test_dates)
     ]
-
-
-def find_cdr_function_from_test_data(
-    test_detect_params, iso3: str, region: str, year: int
-) -> Callable:
-    """
-    Sort out case detection rate from testing numbers, sequentially calling the functions above as required.
-    """
-
-    # Get the testing population denominator
-    testing_pops = get_population_by_agegroup(AGEGROUP_STRATA, iso3, region, year=year)
-
-    # Get the numbers of tests performed
-    test_dates, test_values = get_testing_numbers_for_region(iso3, region)
-
-    # Convert test numbers to per capita testing rates
-    per_capita_tests = [i_tests / sum(testing_pops) for i_tests in test_values]
-
-    # Smooth the testing data if requested
-    if test_detect_params.smoothing_period:
-        smoothed_per_capita_tests = apply_moving_average(
-            per_capita_tests, test_detect_params.smoothing_period
-        )
-    else:
-        smoothed_per_capita_tests = per_capita_tests
-
-    # Scale testing with a time-variant request parameter
-    if test_detect_params.test_multiplier:
-        smoothed_inflated_per_capita_tests = inflate_test_data(
-            test_detect_params.test_multiplier, test_dates, smoothed_per_capita_tests
-        )
-    else:
-        smoothed_inflated_per_capita_tests = smoothed_per_capita_tests
-
-    assert all((val >= 0.0 for val in smoothed_inflated_per_capita_tests))
-
-    # Calculate CDRs and the resulting CDR function
-    cdr_from_tests_func: Callable[[Any], float] = create_cdr_function(
-        test_detect_params.assumed_tests_parameter,
-        test_detect_params.assumed_cdr_parameter,
-        test_detect_params.floor_value,
-    )
-
-    # Get the final CDR function
-    cdr_function = scale_up_function(
-        test_dates,
-        [
-            cdr_from_tests_func(i_test_rate)
-            for i_test_rate in smoothed_inflated_per_capita_tests
-        ],
-        smoothness=0.2,
-        method=4,
-        bound_low=0.0,
-    )
-
-    return cdr_function
